+++ src/Constants.ts
@@ -1,0 +1,6 @@
+export class Constants {
+    public static readonly LogLanguageId: string = 'log';
+
+    public static readonly ContextNameIsShowingProgressIndicators = 'logFileHighlighter.isShowingProgressIndicators';
+}
+